server/requirements.txt
pydantic
pydantic-settings
fastapi
uvicorn
fastapi-mail
jinja2
sqlmodel
sqlalchemy
pyjwt
bcrypt
alembic
redis
celery
asgiref
flower
asyncpg
python-multipart
python-dotenv
itsdangerous
colorama
qrcode
Pillow
pytest
pytest-asyncio
pytest-cov
httpx==0.26.0
aiosqlite
